+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **[BREAKING]** `pull`/`push`/`utils` entries where moved out of the `src` directory. So, import `undercut/push.js` instead of `undercut/src/push.js`. All entry points in the package `root` are public API, but files in the `src` are private and may change their exports, import them at your own risk.
 - **[BREAKING]** `undercut` entry point is now identical to `pull.js`, to use utils you manually need to import `undercut/utils.js`.
+- **[BREAKING]** renamed `makeUnclosable` into `makeObserverUnclosable`.
 
 ## [0.4.0] - 2019-10-15